feat: merge UpdateTargetValidatorLimits and UpdateTargetValidatorLimitsV2
