+++ abi/dvt/UpdateTargetValidatorLimits.abi.json
@@ -39,9 +39,9 @@
             "type": "uint256"
           },
           {
-            "internalType": "bool",
-            "name": "isTargetLimitActive",
-            "type": "bool"
+            "internalType": "uint256",
+            "name": "targetLimitMode",
+            "type": "uint256"
           },
           {
             "internalType": "uint256",
