--- abi/dvt/UpdateTargetValidatorLimitsV2.abi.json
@@ -1,80 +1,0 @@
-[
-  {
-    "inputs": [
-      {
-        "internalType": "address",
-        "name": "_trustedCaller",
-        "type": "address"
-      },
-      {
-        "internalType": "address",
-        "name": "_nodeOperatorsRegistry",
-        "type": "address"
-      }
-    ],
-    "stateMutability": "nonpayable",
-    "type": "constructor"
-  },
-  {
-    "inputs": [
-      { "internalType": "address", "name": "_creator", "type": "address" },
-      { "internalType": "bytes", "name": "_evmScriptCallData", "type": "bytes" }
-    ],
-    "name": "createEVMScript",
-    "outputs": [{ "internalType": "bytes", "name": "", "type": "bytes" }],
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      { "internalType": "bytes", "name": "_evmScriptCallData", "type": "bytes" }
-    ],
-    "name": "decodeEVMScriptCallData",
-    "outputs": [
-      {
-        "components": [
-          {
-            "internalType": "uint256",
-            "name": "nodeOperatorId",
-            "type": "uint256"
-          },
-          {
-            "internalType": "uint256",
-            "name": "targetLimitMode",
-            "type": "uint256"
-          },
-          {
-            "internalType": "uint256",
-            "name": "targetLimit",
-            "type": "uint256"
-          }
-        ],
-        "internalType": "struct UpdateTargetValidatorLimits.TargetValidatorsLimit[]",
-        "name": "",
-        "type": "tuple[]"
-      }
-    ],
-    "stateMutability": "pure",
-    "type": "function"
-  },
-  {
-    "inputs": [],
-    "name": "nodeOperatorsRegistry",
-    "outputs": [
-      {
-        "internalType": "contract INodeOperatorsRegistry",
-        "name": "",
-        "type": "address"
-      }
-    ],
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "inputs": [],
-    "name": "trustedCaller",
-    "outputs": [{ "internalType": "address", "name": "", "type": "address" }],
-    "stateMutability": "view",
-    "type": "function"
-  }
-]
