+++ modules/blockChain/contracts.ts
@@ -280,13 +280,8 @@ export const ContractRewardsShareProgramTopUp = createContractHelpers({
   address: EvmAddressesByType[MotionType.RewardsShareProgramTopUp],
 })
 
-// TODO: merge these two once UpdateTargetValidatorLimits v2 is deployed on the Mainnet
 export const ContractSDVTRegistry = createContractHelpers({
   factory: TypeChain.SDVTRegistryAbi__factory,
-  address: CONTRACT_ADDRESSES.SDVTRegistry,
-})
-export const ContractSDVTRegistryV2 = createContractHelpers({
-  factory: TypeChain.SDVTRegistryV2Abi__factory,
   address: CONTRACT_ADDRESSES.SDVTRegistry,
 })
 
@@ -327,14 +322,9 @@ export const ContractSDVTNodeOperatorManagerChange = createContractHelpers({
   address: EvmAddressesByType[MotionType.SDVTNodeOperatorManagerChange],
 })
 
-// TODO: merge these two once v2 is deployed on the Mainnet
 export const ContractSDVTTargetValidatorLimitsUpdate = createContractHelpers({
   factory: TypeChain.UpdateTargetValidatorLimitsAbi__factory,
   address: EvmAddressesByType[MotionType.SDVTTargetValidatorLimitsUpdate],
-})
-export const ContractSDVTTargetValidatorLimitsUpdateV2 = createContractHelpers({
-  factory: TypeChain.UpdateTargetValidatorLimitsV2Abi__factory,
-  address: EvmAddressesByType[MotionType.SDVTTargetValidatorLimitsUpdateV2],
 })
 
 export const ContractSandboxNodeOperatorsRegistry = createContractHelpers({
