+++ modules/motions/hooks/useContractEvmScript.ts
@@ -58,8 +58,6 @@ export const EVM_CONTRACTS = {
     CONTRACTS.ContractSDVTVettedValidatorsLimitsSet,
   [MotionType.SDVTTargetValidatorLimitsUpdate]:
     CONTRACTS.ContractSDVTTargetValidatorLimitsUpdate,
-  [MotionType.SDVTTargetValidatorLimitsUpdateV2]:
-    CONTRACTS.ContractSDVTTargetValidatorLimitsUpdateV2,
   [MotionType.SDVTNodeOperatorRewardAddressesSet]:
     CONTRACTS.ContractSDVTNodeOperatorRewardAddressesSet,
   [MotionType.SDVTNodeOperatorNamesSet]:
