+++ modules/motions/ui/MotionDescription/DescSDVTTargetValidatorLimitsUpdate.tsx
@@ -15,10 +15,15 @@ export function DescSDVTTargetValidatorLimitsUpdate({
       {callData.map((item, index) => {
         const nodeOperatorId = item.nodeOperatorId.toNumber()
         const nodeOperator = nodeOperatorsList?.[nodeOperatorId]
+        const targetLimitMode = item.targetLimitMode.toNumber()
+
+        let targetLimitModeDesc = 'disabled'
+        if (targetLimitMode == 1) targetLimitModeDesc = 'soft'
+        if (targetLimitMode == 2) targetLimitModeDesc = 'boosted exits'
 
         const nodeOperatorName = nodeOperator ? nodeOperator.name : ''
 
-        if (!item.isTargetLimitActive) {
+        if (targetLimitMode == 0) {
           return (
             <div key={nodeOperatorId}>
               Disable target validator limit for Node Operator{' '}
@@ -35,7 +40,7 @@ export function DescSDVTTargetValidatorLimitsUpdate({
             {nodeOperator && isOnChain
               ? `from ${nodeOperator.targetValidatorsCount} `
               : ''}
-            {`to ${item.targetLimit}`}
+            {`to ${item.targetLimit} in ${targetLimitModeDesc} mode`}
             {index === callData.length - 1 ? '.' : '; '}
           </div>
         )
