+++ modules/motions/ui/MotionFormStartNew/Parts/index.ts
@@ -10,7 +10,6 @@ import * as StartNewSDVTNodeOperatorsActivate from './StartNewSDVTNodeOperatorsA
 import * as StartNewSDVTNodeOperatorsDeactivate from './StartNewSDVTNodeOperatorsDeactivate'
 import * as StartNewSDVTVettedValidatorsLimitsSet from './StartNewSDVTVettedValidatorsLimitsSet'
 import * as StartNewSDVTTargetValidatorLimitsUpdate from './StartNewSDVTTargetValidatorLimitsUpdate'
-import * as StartNewSDVTTargetValidatorLimitsUpdateV2 from './StartNewSDVTTargetValidatorLimitsUpdateV2'
 import * as StartNewSDVTNodeOperatorRewardAddressesSet from './StartNewSDVTNodeOperatorRewardAddressesSet'
 import * as StartNewSDVTNodeOperatorNamesSet from './StartNewSDVTNodeOperatorNamesSet'
 import * as StartNewSDVTNodeOperatorManagersChange from './StartNewSDVTNodeOperatorManagersChange'
@@ -81,8 +80,6 @@ export const formParts = {
     StartNewSDVTVettedValidatorsLimitsSet.formParts,
   [MotionTypeForms.SDVTTargetValidatorLimitsUpdate]:
     StartNewSDVTTargetValidatorLimitsUpdate.formParts,
-  [MotionTypeForms.SDVTTargetValidatorLimitsUpdateV2]:
-    StartNewSDVTTargetValidatorLimitsUpdateV2.formParts,
   [MotionTypeForms.SDVTNodeOperatorRewardAddressesSet]:
     StartNewSDVTNodeOperatorRewardAddressesSet.formParts,
   [MotionTypeForms.SDVTNodeOperatorNamesSet]:
